requirements.txt
# MCP（Model Context Protocol）Python SDK
mcp>=0.9.0

# HTTPクライアント（Places API (New)用）
httpx>=0.25.0

# 環境変数管理用ライブラリ
python-dotenv>=1.0.0

